todo_app_v2/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
black==24.2.0
click==8.1.7
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
flake8==7.0.0
isort==5.13.2
Markdown==3.5.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2023.3.post1
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
wheel==0.42.0
